WIP:  #122 Refactor Code on setting up Modals Folder in Components Folder
