+++ components/Modals/ErrorModalContent.tsx
@@ -3,14 +3,14 @@ import {StyleSheet, View, ScrollView} from 'react-native';
 import Clipboard from '@react-native-community/clipboard';
 import {scale, verticalScale} from 'react-native-size-matters';
 import Toast from 'react-native-toast-message';
-import {useModal} from '../hooks/useModal';
-import {cleanTemporaryErrorRecord, errorGroupParser} from '../utils/errors';
-import {sendErrorCrashEmail} from '../utils/sendMail';
-import Button from './Button';
-import Text from './Text';
-import {screenHeight} from '../utils/layout';
-import {AsyncStoredItems} from '../utils/asyncStorageManager';
-import useAsyncStorage from '../hooks/useAsyncStorage';
+import {useModal} from '../../hooks/useModal';
+import {cleanTemporaryErrorRecord, errorGroupParser} from '../../utils/errors';
+import {sendErrorCrashEmail} from '../../utils/sendMail';
+import Button from '../Button';
+import Text from '../Text';
+import {screenHeight} from '../../utils/layout';
+import {AsyncStoredItems} from '../../utils/asyncStorageManager';
+import useAsyncStorage from '../../hooks/useAsyncStorage';
 
 const ErrorModalContent = (props: {
   errorText: string;
